prettier && move combined arguments
